logic for charts complete. logic for dropdown not finished. do we need dropdown?
I am still trying to finish logic for the dropdown menu. Do we need this or can we put everything in chart. The user comments might satisy the req. for user choice.

I will continue to work on the dropdown menu, unless I hear otherwise.
The HTML structure is complete. Now I need to put that html inside the exchange.html so it has the theme style.